requirements.txt
Flask==3.0.3
pytest==8.3.3
pytest-mock==3.14.0
pytesseract
tesseract
pdf2image
Pillow
opencv-python
numpy
python-docx
openpyxl
xlrd
werkzeug
joblib
pathlib
python-dotenv
scikit-learn
pandas
xlwt